+++ src/Audit.EntityFramework/ConfigurationApi/ContextSettingsConfigurator.cs
@@ -3,7 +3,6 @@ using System;
 namespace Audit.EntityFramework.ConfigurationApi
 {
     public class ContextSettingsConfigurator<T> : IContextSettingsConfigurator<T>
-        where T : IAuditDbContext
     {
         public IContextSettingsConfigurator<T> AuditEventType(string eventType)
         {